+++ src/app/go/mercadopago-pro3/route.ts
@@ -2,5 +2,5 @@ import { NextResponse } from "next/server";
 import { siteConfig } from "@/config/site";
 
 export function GET() {
-  return NextResponse.redirect(siteConfig.referralLinks.ton);
+  return NextResponse.redirect(siteConfig.referralLinks.mercadoPagoPointPro3);
 }