+++ src/config/site.ts
@@ -6,6 +6,20 @@ export const siteConfig = {
   url: "https://fiberstack-hub.vercel.app",
   email: "fiberstacksolucoesoficial@gmail.com",
   creator: "FiberStack Soluções",
+  referralLinks: {
+    ton: "https://ton.com.br/catalogo/?referrer=A06BEA91-498C-4155-9446-C0C4A98C41A6&userAnticipation=0&utm_medium=invite_share&utm_source=revendedor",
+    infinitepayMachine: "https://buy.infinitepay.io/smart?rid=",
+    infinitepayTap: "https://infinitepay.onelink.me/IGWD/",
+    mercadoPagoPointPro3: "https://mpago.li/2J8kuun",
+    mercadoPagoApp: "https://mpago.li/2aFA4jv",
+    mercadoPagoPointSmart2: "https://mpago.li/31gvnCc",
+    mercadoPagoPointAir2: "https://mpago.li/1YAwHQ9",
+    mercadoPagoPointMiniNfc2: "https://mpago.li/2aLDpBv",
+    mercadoPagoPointInvite: "https://mpago.li/2uyXBXN",
+    mercadoPagoPointTap: "https://mpago.li/1SAdk6h",
+    mercadoPagoPaymentLink: "https://mpago.la/1F5JVqr",
+    pagBankMachines: "https://pagbank.vc/indica-maquininhas-ab6ed0ed1",
+  },
   tonReferralUrl:
     "https://ton.com.br/catalogo/?referrer=A06BEA91-498C-4155-9446-C0C4A98C41A6&userAnticipation=0&utm_medium=invite_share&utm_source=revendedor",
   socialLinks: {